[Identity] Create IdentityDocumentScanFragment
