+++ identity/src/main/java/com/stripe/android/identity/navigation/DriverLicenseScanFragment.kt
@@ -15,7 +15,7 @@ import com.stripe.android.identity.states.IdentityScanState.ScanType.DL_FRONT
 internal class DriverLicenseScanFragment(
     identityCameraScanViewModelFactory: ViewModelProvider.Factory,
     identityViewModelFactory: ViewModelProvider.Factory
-) : IdentityCameraScanFragment(
+) : IdentityDocumentScanFragment(
     identityCameraScanViewModelFactory,
     identityViewModelFactory
 ) {
